# [RUNTIME][OPENCL] Create command queue for each module