@masahi 2018-10-30
Do you know why std dev are such large values? It's larger than mean inference time itself.