@tqchen 2018-10-30
We need to be careful here, note that we do need the sequential execution even between functions in different modules, would a separate queue for each module breaks that?