@tqchen 2018-10-30
The concept of queue(stream) in cuda should not be tied to the function itself. The OpenCL runtime does have a limitation that we are only tied to pre-created queues.

The module we used in CUDA is that we allow the call of https://github.com/dmlc/tvm/blob/master/src/runtime/cuda/cuda_device_api.cc#L167 to change the threadlocal context of current stream. This enables concurrency in CUDA, maybe we want to adopt a model similar to that one