@kazum 2018-10-30
@masahi I'm not sure the reason but some of module.run() calls in the result take too much time.  The result is similar even when the CUDA runtime is used.  I guess there's a plenty of room for improvement on module concurrency.