@kazum 2018-10-30
@tqchen  Can you share the example where we need the sequential execution between modules?  I was not aware of the case.

I'll rework this PR based on the stream api.  It looks the right way to go, thanks!