@tqchen 2018-10-30
Although there is no exact example, you can imagine that users would load in multiple modules and call functions from each one of them. Usually user won't expect that these functions don't have serial ordering