@tqchen 2018-12-31
close due to inactive status, new PR that keeps order across modules is welcomed